quilici.Codeflix.Catalog/src/quilici.Codeflix.Infra.EF/UnitOfWork.cs
using quilici.Codeflix.Catalog.Application.Interfaces;

namespace quilici.Codeflix.Catalog.Infra.Data.EF
{
    public class UnitOfWork : IUnitOfWork
    {
        private readonly CodeFlixCatalogDbContext _context;

        public UnitOfWork(CodeFlixCatalogDbContext codeFlixCatalogDbContext)
        {
            _context = codeFlixCatalogDbContext;
        }

        public Task CommitAsync(CancellationToken cancellationToken)
        {
            return _context.SaveChangesAsync(cancellationToken);
        }

        public Task RollbackAsync(CancellationToken cancellationToken)
        {
            return Task.CompletedTask;
        }
    }
}
